File can't be opened, redirects me to browser Fixes #5827
